uploadDate is of type Long

+++ app/src/commonMain/kotlin/model/Cluster.kt
@@ -29,7 +29,7 @@ data class Cluster(
     val coordinate: String,
 
     val uploaderSteamIdHash: String? = null,
-    val uploadDate: Int? = null,
+    val uploadDate: Long? = null,
 
     val gameVersion: Int,
